Distribute tests across workers
