+++ Project.toml
@@ -40,7 +40,6 @@ ZygoteTrackerExt = "Tracker"
 AbstractFFTs = "1.3.1"
 ChainRules = "1.72.2"
 ChainRulesCore = "1.25.1"
-ChainRulesTestUtils = "1"
 Colors = "0.12, 0.13"
 DiffRules = "1.4"
 Distances = "0.10"
@@ -59,18 +58,3 @@ Statistics = "1"
 Tracker = "0.2"
 ZygoteRules = "0.2.7"
 julia = "1.10"
-
-[extras]
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
-ChainRulesTestUtils = "cdddcdb0-9152-4a09-a978-84456f9df70a"
-Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
-Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
-Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
-FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
-FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
-PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-Tracker = "9f7883ad-71c0-57eb-9f7f-b5c9e6d3789c"
-
-[targets]
-test = ["ChainRulesTestUtils", "Conda", "CUDA", "Distances", "FFTW", "FiniteDifferences", "PythonCall", "Test"]
