+++ test/Project.toml
@@ -1,0 +1,22 @@
+[deps]
+ReTestItems = "817f1d60-ba6b-4fd5-9520-3cf149f6a823"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+ChainRulesTestUtils = "cdddcdb0-9152-4a09-a978-84456f9df70a"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
+FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
+FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
+PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
+Tracker = "9f7883ad-71c0-57eb-9f7f-b5c9e6d3789c"
+Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
+ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
+ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
+Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
+SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
+AbstractFFTs = "621f4979-c628-5d54-868e-fcf4e3e8185c"
+LogExpFunctions = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
