+++ test/complex_tests.jl
@@ -1,31 +1,31 @@
-using Zygote, Test, LinearAlgebra
+@testitem "complex" begin
+
+using LinearAlgebra
 
 @testset "basic" begin
+    @test gradient(x -> real(abs(x)*exp(im*angle(x))), 10+20im)[1] ≈ 1
+    @test gradient(x -> imag(real(x)+0.3im), 0.3)[1] ≈ 0
+    @test gradient(x -> imag(conj(x)+0.3im), 0.3 + 0im)[1] ≈ -1im
+    @test gradient(x -> imag(conj(x)+0.3im), 0.3)[1] ≈ 0  # projected to zero
+    @test gradient(x -> abs((imag(x)+0.3)), 0.3 + 0im)[1] ≈ 1im
+    @test gradient(x -> abs((imag(x)+0.3)), 0.3)[1] ≈ 0
 
-@test gradient(x -> real(abs(x)*exp(im*angle(x))), 10+20im)[1] ≈ 1
-@test gradient(x -> imag(real(x)+0.3im), 0.3)[1] ≈ 0
-@test gradient(x -> imag(conj(x)+0.3im), 0.3 + 0im)[1] ≈ -1im
-@test gradient(x -> imag(conj(x)+0.3im), 0.3)[1] ≈ 0  # projected to zero
-@test gradient(x -> abs((imag(x)+0.3)), 0.3 + 0im)[1] ≈ 1im
-@test gradient(x -> abs((imag(x)+0.3)), 0.3)[1] ≈ 0
+    @test gradient(a -> real((a*conj(a))), 0.3im)[1] == 0.6im
+    @test gradient(a -> real((a.*conj(a))), 0.3im)[1] == 0.6im
+    @test gradient(a -> real(([a].*conj([a])))[], 0.3im)[1] == 0.6im
+    @test gradient(a -> real(([a].*conj.([a])))[], 0.3im)[1] == 0.6im
+    @test gradient(a -> real.(([a].*conj.([a])))[], 0.3im)[1] == 0.6im
 
-@test gradient(a -> real((a*conj(a))), 0.3im)[1] == 0.6im
-@test gradient(a -> real((a.*conj(a))), 0.3im)[1] == 0.6im
-@test gradient(a -> real(([a].*conj([a])))[], 0.3im)[1] == 0.6im
-@test gradient(a -> real(([a].*conj.([a])))[], 0.3im)[1] == 0.6im
-@test gradient(a -> real.(([a].*conj.([a])))[], 0.3im)[1] == 0.6im
+    @test gradient(x -> norm((im*x) ./ (im)), 2)[1] == 1
+    @test gradient(x -> norm((im) ./ (im*x)), 2)[1] == -1/4
+    @test gradient(x -> real(det(x)), [1 2im; 3im 4])[1] ≈ [4 3im; 2im 1]
+    @test gradient(x -> real(logdet(x)), [1 2im; 3im 4])[1] ≈ [4 3im; 2im 1]/10
+    @test gradient(x -> real(logabsdet(x)[1]), [1 2im; 3im 4])[1] ≈ [4 3im; 2im 1]/10
 
-@test gradient(x -> norm((im*x) ./ (im)), 2)[1] == 1
-@test gradient(x -> norm((im) ./ (im*x)), 2)[1] == -1/4
-@test gradient(x -> real(det(x)), [1 2im; 3im 4])[1] ≈ [4 3im; 2im 1]
-@test gradient(x -> real(logdet(x)), [1 2im; 3im 4])[1] ≈ [4 3im; 2im 1]/10
-@test gradient(x -> real(logabsdet(x)[1]), [1 2im; 3im 4])[1] ≈ [4 3im; 2im 1]/10
-
-# https://github.com/FluxML/Zygote.jl/issues/705
-@test gradient(x -> imag(sum(exp, x)), [1,2,3])[1] ≈ real(im .* exp.(1:3))
-@test gradient(x -> imag(sum(exp, x)), [1+0im,2,3])[1] ≈ im .* exp.(1:3)
-
-end # @testset
+    # https://github.com/FluxML/Zygote.jl/issues/705
+    @test gradient(x -> imag(sum(exp, x)), [1,2,3])[1] ≈ real(im .* exp.(1:3))
+    @test gradient(x -> imag(sum(exp, x)), [1+0im,2,3])[1] ≈ im .* exp.(1:3)
+end
 
 fs_C_to_R = (real,
              imag,
@@ -119,4 +119,6 @@ end
         sum(r .* abs2.(c)) # This would be calling my actual function depending on r and c
     end
     @test Zygote.hessian(fun, collect(1:9)) ≈ [14 0 0 0 0 0 2 0 0; 0 16 0 0 0 0 0 4 0; 0 0 18 0 0 0 0 0 6; 0 0 0 14 0 0 8 0 0; 0 0 0 0 16 0 0 10 0; 0 0 0 0 0 18 0 0 12; 2 0 0 8 0 0 0 0 0; 0 4 0 0 10 0 0 0 0; 0 0 6 0 0 12 0 0 0]
+end
+
 end
