+++ test/interface_tests.jl
@@ -1,3 +1,5 @@
+@testitem "interface" begin
+
 using Zygote: Grads
 
 @testset "Params" begin
@@ -8,7 +10,7 @@ using Zygote: Grads
     @test w ∈ ps
     @test b ∉ ps
   end
-  
+
   @testset "delete!" begin
     w = rand(2,3)
     b = rand(2)
@@ -23,7 +25,7 @@ using Zygote: Grads
     ps = Params([x])
     copy!(ps, [1, 2])
     @test x == [1, 2]
-    
+
     x = [0,0]
     y = [0]
     ps = Params([x, y])
@@ -35,7 +37,7 @@ using Zygote: Grads
     x = [0, 0]
     copy!(x, ps)
     @test x == [1, 2]
-    
+
     ps = Params([[1,2], [3]])
     x = [0, 0, 0]
     copy!(x, ps)
@@ -65,7 +67,7 @@ using Zygote: Grads
     ps1 = Params([x, y])
     ps2 = Params([x, y])
     ps3 = Params([y, x])
-    @test ps1 == ps2 
+    @test ps1 == ps2
     @test ps1 != ps3  # comparison is order dependent
   end
 
@@ -74,7 +76,7 @@ using Zygote: Grads
     ps1 = Params([x, y])
     ps2 = Params([z])
     ps3 = Params([y, z])
-    
+
     ps = union(ps1, ps2)
     @test ps isa Params
     @test issetequal(ps, Set([x,y,z]))
@@ -82,11 +84,11 @@ using Zygote: Grads
     @test ps isa Params
     @test issetequal(ps, Set([x,y,z]))
 
-    ps = intersect(ps1, ps2) 
+    ps = intersect(ps1, ps2)
     @test ps isa Params
     @test issetequal(ps, Set())
-    
-    ps = intersect(ps1, ps3) 
+
+    ps = intersect(ps1, ps3)
     @test ps isa Params
     @test issetequal(ps, Set([y]))
   end
@@ -102,40 +104,40 @@ end
   @testset "algebra" begin
     w, b = rand(2), rand(2)
     x1, x2 = rand(2), rand(2)
-   
-    gs1 = gradient(() -> sum(w .* x1), Params([w])) 
-    gs2 = gradient(() -> sum(w .* x2), Params([w])) 
-  
+
+    gs1 = gradient(() -> sum(w .* x1), Params([w]))
+    gs2 = gradient(() -> sum(w .* x2), Params([w]))
+
     @test .- gs1 isa Grads
-    @test gs1 .- gs2 isa Grads 
+    @test gs1 .- gs2 isa Grads
     @test .+ gs1 isa Grads
-    @test gs1 .+ gs2 isa Grads 
-    @test 2 .* gs1 isa Grads 
+    @test gs1 .+ gs2 isa Grads
+    @test 2 .* gs1 isa Grads
     @test (2 .* gs1)[w] ≈ 2 * gs1[w]
-    @test gs1 .* 2 isa Grads 
-    @test gs1 ./ 2 isa Grads  
-    @test (gs1 .+ gs2)[w] ≈ gs1[w] .+ gs2[w] 
+    @test gs1 .* 2 isa Grads
+    @test gs1 ./ 2 isa Grads
+    @test (gs1 .+ gs2)[w] ≈ gs1[w] .+ gs2[w]
 
     gs12 = gs1 .+ gs2
     gs1 .+= gs2
-    @test gs12[w] ≈ gs1[w] 
+    @test gs12[w] ≈ gs1[w]
 
     gs3 = gradient(() -> sum(w .* x1), Params([w, b])) # grad nothing with respect to b
-    gs4 = gradient(() -> sum(w .* x2 .+ b), Params([w, b])) 
+    gs4 = gradient(() -> sum(w .* x2 .+ b), Params([w, b]))
 
     @test .- gs3 isa Grads
-    @test gs3 .- gs4 isa Grads 
+    @test gs3 .- gs4 isa Grads
     @test .+ gs3 isa Grads
-    @test gs3 .+ gs4 isa Grads 
-    @test 2 .* gs3 isa Grads 
-    @test gs3 .* 2 isa Grads 
-    @test gs3 ./ 2 isa Grads  
+    @test gs3 .+ gs4 isa Grads
+    @test 2 .* gs3 isa Grads
+    @test gs3 .* 2 isa Grads
+    @test gs3 ./ 2 isa Grads
     @test (gs3 .+ gs4)[w] ≈ gs3[w] .+ gs4[w]
-    @test (gs3 .+ gs4)[b] ≈ gs4[b] 
-    
+    @test (gs3 .+ gs4)[b] ≈ gs4[b]
+
     @test gs3 .+ IdDict(w => similar(w), b => similar(b)) isa Grads
     gs3 .+= IdDict(p => randn(size(p)) for p in keys(gs3))
-    @test gs3 isa Grads 
+    @test gs3 isa Grads
 
     @test_throws ArgumentError gs1 .+ gs4
   end
@@ -166,16 +168,16 @@ end
     w = rand(2)
     x1 = rand(2)
     x2 = rand(2)
-    
-    gs1 = gradient(() -> sum(w .* x1), Params([w])) 
-    gs2 = gradient(() -> sum(w .* x2), Params([w])) 
-    
+
+    gs1 = gradient(() -> sum(w .* x1), Params([w]))
+    gs2 = gradient(() -> sum(w .* x2), Params([w]))
+
     @test map(x -> zeros(2), gs1) isa Grads
-    
-    gs11 = map(x -> clamp.(x, -1e-5, 1e-5), gs1) 
+
+    gs11 = map(x -> clamp.(x, -1e-5, 1e-5), gs1)
     @test gs11 isa Grads
-    @test all(abs.(gs11[w]) .<= 1e-5) 
-  
+    @test all(abs.(gs11[w]) .<= 1e-5)
+
     @test (x -> zeros(2)).(gs1) isa Grads
   end
 
@@ -223,12 +225,12 @@ end
   @testset "iteration" begin
     w, b, x = rand(2), rand(2), rand(2)
     ps = Params([w, b])
-    gs = gradient(() -> sum(tanh.(w .* x .+ b)), ps) 
-    
+    gs = gradient(() -> sum(tanh.(w .* x .+ b)), ps)
+
     # value-based iteration
     foreach(x -> clamp!(x, -1e-5, 1e-5), gs)
-    @test all(abs.(gs[w]) .<= 1e-5) 
-    @test all(abs.(gs[b]) .<= 1e-5) 
+    @test all(abs.(gs[w]) .<= 1e-5)
+    @test all(abs.(gs[b]) .<= 1e-5)
   end
 
   @testset "Params nesting" begin
@@ -237,12 +239,12 @@ end
       b::S
       σ::F
     end
-  
+
     (d::Dense)(x) = d.σ.(d.W * x .+ d.b)
     d = Dense(ones(Float32, 3,3), zeros(Float32, 3), identity)
     ps = Zygote.Params([d.W, d.b])
     r = ones(Float32, 3,3)
-  
+
     gs = gradient(ps) do
       p, pb = pullback(ps) do
         sum(d(r))
@@ -250,9 +252,9 @@ end
       g = pb(p)
       sum(g[d.W]) # + sum(g[d.b])
     end
-  
+
     @test gs[d.W] ≈ fill(81f0, (3,3))
-  
+
     # Test L2
     l2g = gradient(ps) do
       sum(sum(x .^ 2) for x in ps)
@@ -268,5 +270,7 @@ end
     @test sgs[d.W] ≈ fill(1.f0, size(d.W))
     @test sgs[d.b] ≈ fill(1.f0, size(d.b))
   end
+
+end
 
 end
