+++ test/structures_tests.jl
@@ -1,4 +1,7 @@
-using Zygote, Test
+@testitem "structures" begin
+
+using Zygote: bufferfrom
+using Base.Threads: @spawn
 
 function tasks1(x)
   ch = Channel(Inf)
@@ -33,8 +36,6 @@ tasks4(x) = fetch(@async x^2)
 tasks5(x) = fetch(schedule(Task(() -> x^2)))
 @test gradient(tasks5, 5) == (10,)
 
-include("threads.jl")
-
 @test Zygote.pullback(Array, [1f0])[1] == [1f0]
 
 @testset "#300" begin
@@ -45,7 +46,7 @@ include("threads.jl")
 end
 
 struct A594 x::Float64 end
-  
+
 @testset "#594" begin
   f(a,v) = a.x + v
   g(A,V) = sum(f.(A,V))
@@ -65,4 +66,38 @@ end
 
   m, b = Zygote._pullback(Zygote.Context(), nameof, M)
   @test b(m) === nothing
+end
+
+@testset "threads" begin
+    function threads1(x)
+      ch = Channel(0)
+      @sync begin
+        @spawn put!(ch, x^2)
+        take!(ch)
+      end
+    end
+
+    @test gradient(threads1, 5) == (10,)
+
+    function threads2(xs)
+      n = length(xs)
+      chunks = view(xs, 1:n÷2), view(xs, n÷2+1:n)
+      p = bufferfrom([0.0, 0.0])
+      @sync begin
+        for i = 1:2
+          @spawn begin
+            s = zero(eltype(chunks[i]))
+            for j = 1:length(chunks[i])
+              s += chunks[i][j]
+            end
+            p[i] = s
+          end
+        end
+      end
+      return p[1]+p[2]
+    end
+
+    @test gradient(threads2, [1, 2, 3, 4]) == ([1, 1, 1, 1],)
+end
+
 end
