+++ test/tools_tests.jl
@@ -1,23 +1,25 @@
+@testitem "tools" begin
+
 @testset "Buffer" begin
   w = rand(2,3)
   b = rand(2)
   c = rand(2)
-  
+
   buff = Zygote.Buffer([])
   @test length(buff) == 0
-  
+
   push!(buff, w)
   push!(buff, b)
   push!(buff, c)
   @test length(buff) == 3
-  @test buff[1] === w 
-  @test buff[2] === b 
-  @test buff[3] === c 
-  
+  @test buff[1] === w
+  @test buff[2] === b
+  @test buff[3] === c
+
   deleteat!(buff, 2)
   @test length(buff) == 2
-  @test buff[1] === w 
-  @test buff[2] === c 
+  @test buff[1] === w
+  @test buff[2] === c
 
   deleteat!(buff, 1)
   @test length(buff) == 1
@@ -60,5 +62,7 @@ end
 
   gs56 = gradient(f56, [1.0])
   @test gs56[1] == [0.]
+
+end
 
 end
