+++ test/utils_tests.jl
@@ -1,4 +1,7 @@
+@testitem "utils" begin
+
 using ForwardDiff
+using LinearAlgebra
 using Zygote: hessian_dual, hessian_reverse
 
 @testset "hessian: $hess" for hess in [hessian_dual, hessian_reverse]
@@ -133,4 +136,6 @@ using ForwardDiff
   @test ForwardDiff.gradient(y->gradf1218(x,y), y) == ones(5)
   # this returns (nothing,) -- now prints a warning
   @test_broken Zygote.gradient(y->gradf1218(x,y), y) == ones(5)
+end
+
 end
